python/requirements.txt
pyzmq==25.1.2
numpy==1.26.3
opencv-python==4.9.0.80
pillow==10.2.0
ffmpeg-python
torch
torchvision
